post deletion
post deletion(fix) and made the username appear on the nav bar

+++ app/Http/Controllers/PostController.php
@@ -8,12 +8,14 @@ use App\Http\Requests;
 
 use App\Post;
 
+use Illuminate\Support\Facades\Auth;
+
 class PostController extends Controller
 {
 
     public function getDashboard() 
     {
-        $posts = Post::all();
+        $posts = Post::latest()->get();
         return view('dashboard',['posts' => $posts]);
     }
 
@@ -37,6 +39,12 @@ class PostController extends Controller
     public function getDeletePost($post_id)
     {
         $post = Post::where('id', $post_id)->first();
+
+        if(Auth::user != $post->user)
+        {
+            return redirect()->back();
+        }
+
         $post->delete();
         return redirect()->route('dashboard')->with(['message' => 'Successfully deleted!']);
     }
